requirements.txt
overrides~=6.1.0
torch~=1.12.0+cu116
numpy~=1.23.1
scikit-learn~=1.1.1
matplotlib~=3.5.2
pandas~=1.4.3
tqdm~=4.64.0
paramiko~=2.11.0
PyYAML~=6.0
latextable~=0.3.0
texttable~=1.6.4
setuptools~=57.0.0
torchvision~=0.13.0+cu116
Pillow~=9.2.0
scipy~=1.8.1
optuna~=2.10.1
gym~=0.25.0
seaborn~=0.11.2
wandb~=0.12.21
